XWIKI-18773: Allow to call ConversionFilter in a specific ResourceReferenceHandler
  * Create a dedicated component to perform request parameter
    conversion, and use it in ConversionFilter
