+++ xwiki-platform-core/xwiki-platform-wysiwyg/xwiki-platform-wysiwyg-api/src/main/resources/META-INF/components.txt
@@ -9,3 +9,4 @@ org.xwiki.wysiwyg.internal.filter.http.MutableHttpServletRequestFactory
 org.xwiki.wysiwyg.internal.importer.OfficeAttachmentImporter
 org.xwiki.wysiwyg.internal.importer.OfficeMacroImporter
 org.xwiki.wysiwyg.script.WysiwygEditorScriptService
+org.xwiki.wysiwyg.internal.converter.DefaultRequestParameterConverter
